+++ src/index.ts
@@ -1,4 +1,6 @@
+import SernEmitter from './handler/sernEmitter';
 export * as Sern from './handler/sern';
 export * from './types/handler';
 export * from './handler/structures/structxports';
 export * from './handler/plugins/plugin';
+export { SernEmitter };
